full excersises

+++ Part4/bloglist/controllers/blogs.js
@@ -1,10 +1,12 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
 
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 
 })
@@ -16,9 +18,10 @@ blogsRouter.get('/:id', async (request, response) => {
 
 })
 
-blogsRouter.post('/', async (request, response) => {
 
-    const user = await User.findById(request.body.userId)
+blogsRouter.post('/', middleware.tokenExtractor, middleware.userExtractor, async (request, response) => {
+
+    const user = request.user
 
     const blog = new Blog({
         title: request.body.title,
@@ -50,13 +53,20 @@ blogsRouter.put('/:id', async (request, response) => {
 
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.tokenExtractor, middleware.userExtractor, async (request, response) => {
 
-    console.log(request.params.id)
+    const user = request.user
+    
+    const blog = await Blog.findById(request.params.id)
 
-    const deletedBlog = await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    if (blog.user.toString() === user._id.toString()) {
+        const deletedBlog = await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'Cant delete other people posts' })
+    }
 
+    
 })
